# docs/swagger.yaml
basePath: /v1
definitions:
  controllers.ServiceMeta:
    properties:
      address:
        type: string
      ext:
        type: object
      id:
        type: string
      metricsAddress:
        type: string
      metricsType:
        type: string
      outAddress:
        type: string
      path:
        type: string
      status:
        type: boolean
      typ:
        type: string
      version:
        type: string
    type: object
...
  models.User:
    properties:
      createTime:
        description: 创建时间
        type: integer
      id:
        description: 用户ID
        type: string
      ids:
        description: Ids
        items:
          type: string
        type: array
      password:
        description: 登陆密码
        type: string
      realName:
        description: 用户姓名
        type: string
      roles:
        description: 用户角色
        items:
          type: string
        type: array
      salt:
        description: 密码加密
        type: string
      username:
        description: 登陆名
        type: string
    required:
    - password
    - username
    type: object
  resp.Response:
    properties:
      code:
        description: 错误码 状态码不为200时返回
        type: integer
      data:
        description: 返回数据
        type: object
      message:
        description: 错误信息 状态码不为200时返回
        type: string
      more_info:
        description: 错误扩展信息 状态码不为200时返回
        type: object
      status:
        description: 状态码 200:成功   其他失败
        type: integer
      type:
        description: 错误类型 状态码不为200时返回
        type: string
    type: object
info:
  contact:
    email: leizhou.lin@watrix.ai
    name: API Support
    url: github.com/souliot
  description: Golang api of demo
  termsOfService: github.com/souliot
  title: DevOps 开发文档
  version: v1.0.0
paths:
...
  /service/{env}/{path}/{type}/{id}:
    delete:
      consumes:
      - application/json
      description: 删除离线服务节点
      parameters:
      - description: service env
        in: path
        name: env
        required: true
        type: string
      - description: service path
        in: path
        name: path
        required: true
        type: string
      - description: service type
        in: path
        name: type
        required: true
        type: string
      - description: service id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      summary: 删除离线服务节点
      tags:
      - 服务
  /service/all:
    get:
      consumes:
      - application/json
      description: 获取服务列表
      parameters:
      - description: service env
        in: query
        name: env
        type: string
      - description: service path
        in: query
        name: path
        type: string
      - description: service type
        in: query
        name: type
        type: string
      - description: service id
        in: query
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 返回数据 ServiceMeta
          schema:
            allOf:
            - $ref: '#/definitions/resp.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/controllers.ServiceMeta'
                  type: array
              type: object
      summary: 获取服务列表
      tags:
      - 服务
  /user:
    get:
      consumes:
      - application/json
      description: 获取用户列表
      parameters:
      - description: node address
        in: query
        name: name
        type: string
      - description: node address
        in: query
        name: userName
        type: string
      - description: page
        in: query
        name: page
        type: string
      - description: pageSize
        in: query
        name: pageSize
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 返回数据 []User
          schema:
            allOf:
            - $ref: '#/definitions/resp.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.User'
                  type: array
              type: object
      summary: 获取用户列表
      tags:
      - 用户
    post:
      consumes:
      - application/json
      description: 添加用户的接口
      parameters:
      - description: 用户
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: 返回数据 User
          schema:
            allOf:
            - $ref: '#/definitions/resp.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.User'
              type: object
      summary: 添加用户
      tags:
      - 用户
    put:
      consumes:
      - application/json
      description: 修改用户的接口
      parameters:
      - description: 用户
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: 返回数据 User
          schema:
            allOf:
            - $ref: '#/definitions/resp.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.User'
              type: object
      summary: 修改用户
      tags:
      - 用户
  /user/{id}:
    delete:
      consumes:
      - application/json
      description: 删除用户的接口
      parameters:
      - description: node id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 返回数据 User
          schema:
            allOf:
            - $ref: '#/definitions/resp.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.User'
              type: object
      summary: 删除用户
      tags:
      - 用户
    get:
      consumes:
      - application/json
      description: 获取用户信息
      parameters:
      - description: node id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 返回数据 User
          schema:
            allOf:
            - $ref: '#/definitions/resp.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.User'
              type: object
      summary: 获取用户信息
      tags:
      - 用户
  /user/getUserInfo:
    get:
      consumes:
      - application/json
      description: 获取登录用户信息的接口
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 返回数据 User
          schema:
            allOf:
            - $ref: '#/definitions/resp.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.User'
              type: object
      summary: 获取登录用户信息
      tags:
      - 用户
  /user/login:
    post:
      consumes:
      - application/json
      description: 用户登录的接口
      parameters:
      - description: 用户
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: 返回数据 User
          schema:
            allOf:
            - $ref: '#/definitions/resp.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.User'
              type: object
      summary: 用户登录
      tags:
      - 用户
  /user/logout:
    post:
      consumes:
      - application/json
      description: 用户登出的接口
      parameters:
      - description: 用户
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: 返回数据 User
          schema:
            $ref: '#/definitions/resp.Response'
      summary: 用户登出
      tags:
      - 用户
swagger: "2.0"